# storage/scribe/openapi.yaml
openapi: 3.0.3
info:
  title: 'Evolução local de snipe-it API Documentation'
  description: 'This documentation aims to provide all the information you need to work with the Snipe-IT API.'
  version: 1.0.0
servers:
  -
    url: 'https://snipe-it.test'
...
paths:
...
  '/api/v1/categories/{item_type}/selectlist':
    get:
      summary: Selectlist
      operationId: selectlist
      description: ''
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: 'Unauthorized or unauthenticated.'
                properties:
                  error:
                    type: string
                    example: 'Unauthorized or unauthenticated.'
      tags:
        - Categories
    parameters:
      -
        in: path
        name: item_type
        description: ''
        example: architecto
        required: true
        schema:
          type: string
...
  /api/v1/suppliers:
    get:
      summary: 'List Suppliers'
      operationId: listSuppliers
      description: ''
      parameters:
        -
          in: query
          name: search
          description: 'Search term to filter results.'
          example: Acme
          required: false
          schema:
            type: string
            description: 'Search term to filter results.'
            example: Acme
            nullable: false
        -
          in: query
          name: name
          description: 'Filter by exact supplier name.'
          example: 'Acme Corp'
          required: false
          schema:
            type: string
            description: 'Filter by exact supplier name.'
            example: 'Acme Corp'
            nullable: false
        -
          in: query
          name: address
          description: 'Filter by exact address.'
          example: '123 Main St'
          required: false
          schema:
            type: string
            description: 'Filter by exact address.'
            example: '123 Main St'
            nullable: false
        -
          in: query
          name: address2
          description: 'Filter by exact address2.'
          example: 'Suite 100'
          required: false
          schema:
            type: string
            description: 'Filter by exact address2.'
            example: 'Suite 100'
            nullable: false
        -
          in: query
          name: city
          description: 'Filter by exact city.'
          example: Springfield
          required: false
          schema:
            type: string
            description: 'Filter by exact city.'
            example: Springfield
            nullable: false
        -
          in: query
          name: state
          description: 'Filter by exact state.'
          example: IL
          required: false
          schema:
            type: string
            description: 'Filter by exact state.'
            example: IL
            nullable: false
        -
          in: query
          name: zip
          description: 'Filter by exact zip code.'
          example: '62701'
          required: false
          schema:
            type: string
            description: 'Filter by exact zip code.'
            example: '62701'
            nullable: false
        -
          in: query
          name: country
          description: 'Filter by exact country.'
          example: USA
          required: false
          schema:
            type: string
            description: 'Filter by exact country.'
            example: USA
            nullable: false
        -
          in: query
          name: phone
          description: 'Filter by exact phone number.'
          example: 555-1234
          required: false
          schema:
            type: string
            description: 'Filter by exact phone number.'
            example: 555-1234
            nullable: false
        -
          in: query
          name: fax
          description: 'Filter by exact fax number.'
          example: 555-5678
          required: false
          schema:
            type: string
            description: 'Filter by exact fax number.'
            example: 555-5678
            nullable: false
        -
          in: query
          name: email
          description: 'Filter by exact email address.'
          example: info@example.org
          required: false
          schema:
            type: string
            description: 'Filter by exact email address.'
            example: info@example.org
            nullable: false
        -
          in: query
          name: url
          description: 'Filter by exact URL.'
          example: 'http://example.com'
          required: false
          schema:
            type: string
            description: 'Filter by exact URL.'
            example: 'http://example.com'
            nullable: false
        -
          in: query
          name: notes
          description: 'Filter by exact notes.'
          example: 'This is a note.'
          required: false
          schema:
            type: string
            description: 'Filter by exact notes.'
            example: 'This is a note.'
            nullable: false
        -
          in: query
          name: sort
          description: 'Column to sort results by. Allowed values: id, name, address, address2, city, state, country, zip, phone, contact, fax, email, image, assets_count, licenses_count, accessories_count, components_count, consumables_count, url, notes. Default: created_at.'
          example: name
          required: false
          schema:
            type: string
            description: 'Column to sort results by. Allowed values: id, name, address, address2, city, state, country, zip, phone, contact, fax, email, image, assets_count, licenses_count, accessories_count, components_count, consumables_count, url, notes. Default: created_at.'
            example: name
            nullable: false
        -
          in: query
          name: order
          description: 'Order of sorted results. Allowed values: asc, desc. Default: desc.'
          example: asc
          required: false
          schema:
            type: string
            description: 'Order of sorted results. Allowed values: asc, desc. Default: desc.'
            example: asc
            nullable: false
        -
          in: query
          name: offset
          description: 'Offset/starting position of the results. Default: 0.'
          example: 0
          required: false
          schema:
            type: integer
            description: 'Offset/starting position of the results. Default: 0.'
            example: 0
            nullable: false
        -
          in: query
          name: limit
          description: 'Limit the number of results returned. Default: 25. Maximum: 100.'
          example: 50
          required: false
          schema:
            type: integer
            description: 'Limit the number of results returned. Default: 25. Maximum: 100.'
            example: 50
            nullable: false
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: 'Unauthorized or unauthenticated.'
                properties:
                  error:
                    type: string
                    example: 'Unauthorized or unauthenticated.'
      tags:
        - Suppliers
    post:
      summary: 'Create Supplier'
      operationId: createSupplier
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Suppliers
...